Getting more optimized

+++ contracts/Geometry.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.4.23;
 library Geometry {
 
     struct Rect {
-        uint16 x;
-        uint16 y;
-        uint16 w;
-        uint16 h;
+        uint256 x;
+        uint256 y;
+        uint256 w;
+        uint256 h;
     }
 
     function doRectanglesOverlap(Rect memory a, Rect memory b) internal pure returns (bool) {
@@ -22,8 +22,8 @@ library Geometry {
         result.y = a.y > b.y ? a.y : b.y;
 
         // Take the lesser of the x2 and y2 values
-        uint16 resultX2 = a.x + a.w < b.x + b.w ? a.x + a.w : b.x + b.w;
-        uint16 resultY2 = a.y + a.h < b.y + b.h ? a.y + a.h : b.y + b.h;
+        uint256 resultX2 = a.x + a.w < b.x + b.w ? a.x + a.w : b.x + b.w;
+        uint256 resultY2 = a.y + a.h < b.y + b.h ? a.y + a.h : b.y + b.h;
 
         // Set our width and height here
         result.w = resultX2 - result.x;
